move grandpa integration code
